backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
PyJWT==2.8.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0 # For production deployment
openai==1.3.7 # For AI model interactions
supabase-py==2.4.4 # Supabase Python client
